http_wol_trigger/requirements.txt
fastapi
paho-mqtt
python-dotenv